pyproject.toml
[tool.poetry]
name = "zero_hunger"
version = "0.1.0"
description = "A zero hunger solution by solving farmers problems and agricultural challenges"
authors = [
    "Wadinga Leonard N.",
    "Sone Gillis",
    "Gaston Che",
    "Crespy",
    "Valery Yola",
    "HOIBI HOIBI",
    "Sonia",
]


[tool.poetry.dependencies]
python = "^3.7, <3.9"
django = "^3.0"
djangorestframework = "^3.11.0"
python-decouple = "^3.3"
django-cors-headers = "^3.2.1"
drf-yasg = "^1.17.1"
django-rest-swagger = "^2.2.0"
dj-rest-auth = "^1.0.5"
djangorestframework-simplejwt = "^4.4.0"
Pillow = "^7.1.2"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.4"
pytest-django = "^3.9.0"
flake8 = "^3.7.9"
flake8-django = "^1.0.0"
black = "^19.10b0"
isort = "^4.3.21"
pytest-ordering = "^0.6"


[tool.poetry.scripts]
app = 'manage:wsgi'


[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
include_trailing_space = true


[tool.flake8]
ignore = ['E501', 'E231']
max-complexity = 10
exclude = '''
  .pytest_cache
  .DS_Store
  __pycache__
  **/**/__pycache__
  *.pyc
  .git
  setup.py
  build
  dist
  releases
  .venv
  .tox
  .vscode
  .github
  tests/fixtures/
  tests/utils/fixtures/
  .env
  venv
'''


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

